+++ crates/picodroid-core/src/graphics/lvgl/widgets/mod.rs
@@ -9,6 +9,7 @@
 pub mod alert_dialog;
 pub mod button;
 pub mod check_box;
+pub mod circular_progress_indicator;
 pub mod date_picker;
 pub mod edit_text;
 pub mod frame_layout;